added game view
