+++ src/tower/__main__.py
@@ -2,56 +2,43 @@
 Platformer Game
 """
 
-from pathlib import Path
-
 import arcade
 
 # Constants
+from tower.src.tower.game_view import GameView
+
 SCREEN_WIDTH = 960
 SCREEN_HEIGHT = 640
 SCREEN_TITLE = "Platformer"
 
 
-class MyGame(arcade.Window):
-    """
-    Main application class.
-    """
+class StartView(arcade.View):
+    def on_show_view(self):
+        """ This is run once when we switch to this view """
+        arcade.set_background_color(arcade.csscolor.BLACK)
 
-    def __init__(self):
-        # Call the parent class and set up the window
-        super().__init__(SCREEN_WIDTH, SCREEN_HEIGHT, SCREEN_TITLE)
-
-        arcade.set_background_color(arcade.csscolor.CORNFLOWER_BLUE)
-
-        self.tile_map = None
-        self.scene = None
-
-        self.towers = None
-        self.enemies = None
-
-    def setup(self):
-        """Set up the game here. Call this function to restart the game."""
-
-        level_1_path = Path(__file__).parent.parent.parent / "assets" / "levels" / "level_1.tmx"
-        self.tile_map = arcade.load_tilemap(level_1_path)
-
-        self.scene = arcade.Scene.from_tilemap(self.tile_map)
-
-        self.towers = arcade.SpriteList()
-        self.enemies = arcade.SpriteList()
+        # to reset the viewport back to the start so we can see what we draw.
+        arcade.set_viewport(0, self.window.width, 0, self.window.height)
 
     def on_draw(self):
-        """Render the screen."""
-
         self.clear()
+        arcade.draw_text("Welcome to our beautiful game", self.window.width / 2, self.window.height / 2,
+                         arcade.color.WHITE, font_size=50, anchor_x="center")
+        arcade.draw_text("Click to start a game", self.window.width / 2, self.window.height / 2 - 75,
+                         arcade.color.WHITE, font_size=20, anchor_x="center")
 
-        self.scene.draw()
+    def on_mouse_press(self, _x, _y, _button, _modifiers):
+        """ If the user presses the mouse button, start the game. """
+        game_view = GameView()
+        game_view.setup()
+        self.window.show_view(game_view)
 
 
 def main():
     """Main function"""
-    window = MyGame()
-    window.setup()
+    window = arcade.Window(SCREEN_WIDTH, SCREEN_HEIGHT, SCREEN_TITLE)
+    start_view = StartView()
+    window.show_view(start_view)
     arcade.run()
